Save improvements, Added mistake protection
Save improvements: Added "Save as ..." button, it let choose the save location. Now the "Save" button has 2 different behaviour, based on whether the project has already been saved or not. In the first case, it save the project in the same file; in the second case it let choose the file location.
Mistake protection: If the project has been modified since the last save, when "New project", "Load Project" are pressed or the application is closed, it will notify that there are not-saved changes and it will asks to continue or not.